refactor: getLinkTargets belongs with the other link rules
The shapes a source may connect to is a question about links, answered by
`validateConnection()` — the same rule the pointer path enforces. It sat as
a private function in the `QuickLink` component, where nothing else could
ask it.

It moves to `utils/links.ts`, next to `resolveDefaultLinkType` and
`validateAndReplaceConnections`, taking the ordering with it: a pool or an
artifact reads last because it is the rarer choice, and that is part of the
answer rather than the picker's own presentation.

`describe()` stays in the component — reading a shape's name for a row is
presentation, not a link rule.
